requirements.txt
Flask
Flask-CORS
werkzeug
Pillow
pdf2image
opencv-python
numpy
EasyOCR
gunicorn